examples/VoiceTranslation/requirements.txt
braintrust
python-dotenv
requests
openai==0.28.0
pydantic
